requirements.txt
paho-mqtt
Flask
pyyaml
#adafruit-blinka
#adafruit-circuitpython-mcp230xx
RPi.GPIO
mcp23017
smbus2